Move base entity mock creator to the mocks directory
